+++ Core/Matcher/ObjectStateMatcher.php
@@ -11,11 +11,11 @@ class ObjectStateMatcher extends AbstractMatcher implements KeyMatcherInterface
 {
     use FlexibleKeyMatcherTrait;
 
-    const MATCH_STATE_ID = 'objectstate_id';
-    const MATCH_STATE_IDENTIFIER = 'objectstate_identifier';
+    const MATCH_OBJECTSTATE_ID = 'objectstate_id';
+    const MATCH_OBJECTSTATE_IDENTIFIER = 'objectstate_identifier';
 
     protected $allowedConditions = array(
-        self::MATCH_STATE_ID, self::MATCH_STATE_IDENTIFIER,
+        self::MATCH_OBJECTSTATE_ID, self::MATCH_OBJECTSTATE_IDENTIFIER,
         // aliases
         'id', 'identifier'
     );
@@ -46,11 +46,11 @@ class ObjectStateMatcher extends AbstractMatcher implements KeyMatcherInterface
 
             switch ($key) {
                 case 'id':
-                case self::MATCH_STATE_ID:
+                case self::MATCH_OBJECTSTATE_ID:
                    return new ObjectStateCollection($this->findObjectStatesById($values));
 
                 case 'identifier':
-                case self::MATCH_STATE_IDENTIFIER:
+                case self::MATCH_OBJECTSTATE_IDENTIFIER:
                     return new ObjectStateCollection($this->findObjectStatesByIdentifier($values));
             }
         }
@@ -59,54 +59,77 @@ class ObjectStateMatcher extends AbstractMatcher implements KeyMatcherInterface
     protected function getConditionsFromKey($key)
     {
         if (is_int($key) || ctype_digit($key)) {
-            return array(self::MATCH_STATE_ID => $key);
+            return array(self::MATCH_OBJECTSTATE_ID => $key);
         }
-        return array(self::MATCH_STATE_IDENTIFIER => $key);
+        return array(self::MATCH_OBJECTSTATE_IDENTIFIER => $key);
     }
 
     /**
-     * @param int[] $stateIds
+     * @param int[] $objectStateIds
      * @return ObjectState[]
      */
-    protected function findObjectStatesById(array $stateIds)
+    protected function findObjectStatesById(array $objectStateIds)
     {
-        $states = [];
+        $objectStates = [];
 
-        foreach ($stateIds as $stateId) {
+        foreach ($objectStateIds as $objectStateId) {
             // return unique contents
-            $state = $this->repository->getObjectStateService()->loadObjectState($stateId);
-            $states[$state->id] = $state;
+            $objectState = $this->repository->getObjectStateService()->loadObjectState($objectStateId);
+            $objectStates[$objectState->id] = $objectState;
         }
 
-        return $states;
+        return $objectStates;
     }
 
     /**
-     * @param string[] $stateIdentifiers
+     * @param string[] $stateIdentifiers Accepts the state identifier if unique, otherwise "group-identifier/state-identifier"
      * @return ObjectState[]
      * @throws NotFoundException
      */
     protected function findObjectStatesByIdentifier(array $stateIdentifiers)
     {
         // we have to build this list, as the ObjectStateService does not allow to load a State by identifier...
-        $statesList = array();
-        $groups = $this->repository->getObjectStateService()->loadObjectStateGroups();
-        foreach ($groups as $group) {
-            $groupStates = $this->repository->getObjectStateService()->loadObjectStates($group);
-            foreach($groupStates as $groupState) {
-                $statesList[$groupState->identifier] = $groupState;
-            }
-        }
+        $statesList = $this->loadAvailableStates();
 
         $states = [];
 
         foreach ($stateIdentifiers as $stateIdentifier) {
             if (!isset($statesList[$stateIdentifier])) {
-                throw new NotFoundException("ObjectState", $stateIdentifier);
+                // a quick and dirty way of letting the user know that he/she might be using a non-unique identifier
+                throw new NotFoundException("ObjectState", $stateIdentifier . ' missing or non unique');
             }
             $states[$statesList[$stateIdentifier]->id] = $statesList[$stateIdentifier];
         }
 
         return $states;
+    }
+
+    /**
+     * @return ObjectState[] key: the state identifier (for unique identifiers), group_identifier/state_identifier for all
+     */
+    protected function loadAvailableStates()
+    {
+        $statesList = array();
+        $nonuniqueIdentifiers = array();
+
+        $groups = $this->repository->getObjectStateService()->loadObjectStateGroups();
+        foreach ($groups as $group) {
+            $groupStates = $this->repository->getObjectStateService()->loadObjectStates($group);
+            foreach($groupStates as $groupState) {
+                // we always add the states using 'group/state' identifiers
+                $statesList[$group->identifier.'/'.$groupState->identifier] = $groupState;
+                // we only add the state using plain identifier if it is unique
+                if (isset($statesList[$groupState->identifier])) {
+                    unset($statesList[$groupState->identifier]);
+                    $nonuniqueIdentifiers[] = $groupState->identifier;
+                } else {
+                    if (!isset($nonuniqueIdentifiers[$groupState->identifier])) {
+                        $statesList[$groupState->identifier] = $groupState;
+                    }
+                }
+            }
+        }
+
+        return $statesList;
     }
 }